3.4-django-testing/django_testing/requirements.txt
django~=4.1
djangorestframework~=3.13.1
django-filter~=22.1
psycopg2-binary
python-dotenv~=0.21.0

